Update ppcg & prl for passing work_size instead of grid_size
grid_size = ceild(work_size, block_size)

There was an error when having more work dims than grid dims, in that it returned the block size (= least number of blocks) instead of 1, ending up enqueing empty work (the generated kernels have out-of-work checks)
To avoid more changes to mainline ppcg, pass work_items as for --target=openml.

Because this is a API break, update ppcg and prl with this change at the same time.